backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai==1.*
pydantic
pytz


